requirements.txt
langchain-openai
langchain-core
langgraph
pydantic
python-dotenv
IPython
ddgs
bs4
html2text